User avatar updates when necessary. Implement cache busting option for avatar image hook.
